backend/requirements.txt
Django==4.1.2
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.22.1
gunicorn==20.1.0
psycopg2==2.9.2
celery==5.2.7
redis==4.3.3
